stack-docker: make the JVM get pid 1 to allow graceful shutdown
The "sh -c" entry point is nice to work around the Docker var replacement limitation in CMD statements.
However the shell keeps running along with the JVM and the shell gets pid 1.
Consequently, when you run `docker stop`, the shell process gets the SIGTERM signal, but does not forward it to the JVM.
This means Vert.x does not shutdown gracefully.